backend/pyproject.toml
[project]
name = "fastmatch"
version = "0.1.0"
description = ""
authors = [
    {name = "Amirhan Botagariev",email = "amirkhan1botagariev@mail.ru"}
]
license = {text = "MIT"}
package-mode = false
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-magic (>=0.4.27,<0.5.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "boto3 (>=1.41.2,<2.0.0)",
    "minio (>=7.2.19,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "google-genai (>=1.52.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
